requirements.txt
tweepy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
slack-sdk
python-dotenv
